don't let panels revert to display: none
if the active panel's "display: block" element css override is removed when the panel is closed, it goes back to .panel's "display: none", the rippleWave element also becomes display: none;, and thus the transitionend event that does .remove() on the rippleWave element.
so either we find another way to remove() without relying on the transitionend event, or we just keep it at display: block; all the time.