Fix walk status bug.
Bug caused by deferEvaluation option of computed observable being set to true in order to avoid an 'activeDog is undefined' error when the ViewModel is first instantiated (computed observables try to evaluate their value the moment their parent is created. at that point of execution, there is no activeDog set yet).

walkStatus is now implemented with a plain Javascript function.