Fix error when applying bindings to 2nd instance
of page. Instead of hard-coding to the first element found with the data-page="...", use the Page Data object's container property to grab the correct HTMLElement directly.

Now, when a duplicate page is loaded (either from Home > Menu > Home or by refreshing with "#!/index.html" as the hash state, the bindings are applied correctly to the duplicate page instead of the first one found.